修改判断

+++ install/install.sh
@@ -39,14 +39,21 @@ mkdir ./templates/static
 mkdir ./config
 
 echo "Create configuration file..."
-
-cp -i ./example/menu.example.json ./config/menu.json
-cp -i ./example/page.example.json ./config/page.json
-cp -i ./example/system.example.json ./config/system.json
-
-cp -i ./example/start.example.json ./start.json
-cp -i ./example/uwsgi.example.json ./uwsgi.json
+if [ ! -f "./config/menu.json" ]; then
+    cp -i ./example/menu.example.json ./config/menu.json
+fi
+if [ ! -f "./config/page.json" ]; then
+    cp -i ./example/page.example.json ./config/page.json
+fi
+if [ ! -f "./config/system.json" ]; then
+    cp -i ./example/system.example.json ./config/system.json
+fi
+if [ ! -f "./start.json" ]; then
+    cp -i ./example/start.example.json ./start.json
+fi
+if [ ! -f "./uwsgi.json" ]; then
+    cp -i ./example/uwsgi.example.json ./uwsgi.json
+fi
 
 chmod +x manage.py
-
 echo "The installation is complete! Please edit $(pwd)/config/system.json file."
